test_agent.py
"""
Test Customer Success Copilot Agent
Verify everything works before building the demo interface
"""

import os
from dotenv import load_dotenv
from weave_integration import WeaveTrackedCustomerSuccessCopilot as CustomerSuccessCopilot
from sample_data import SAMPLE_EMAILS, get_demo_email

load_dotenv()

def test_agent_analysis():
    """Test the agent with a realistic customer email"""
    
    # Initialize agent
    api_key = os.getenv("OPENAI_API_KEY")
    if not api_key:
        print("❌ Error: OPENAI_API_KEY not found")
        return False
    
    agent = CustomerSuccessCopilot(api_key)
    print("✅ Customer Success Copilot initialized")
    
    # Test with frustrated premium customer
    test_case = SAMPLE_EMAILS["frustrated_premium"]
    email_content = test_case["email"]
    customer_profile = test_case["customer_profile"]
    
    print(f"\n🔍 Testing Analysis...")
    print(f"Customer: {customer_profile.name} ({customer_profile.tier} tier)")
    print(f"Email length: {len(email_content)} characters")
    
    # Run analysis
    analysis = agent.analyze_customer_communication(email_content, customer_profile)
    
    # Display results
    print(f"\n📊 ANALYSIS RESULTS:")
    print("=" * 50)
    
    print(f"💭 Sentiment: {analysis.sentiment['label']} (confidence: {analysis.sentiment['confidence']:.2f})")
    print(f"⚡ Urgency: {analysis.urgency['level']}")
    print(f"   Reasoning: {analysis.urgency['reasoning']}")
    
    print(f"\n🎯 Customer Intent:")
    for intent in analysis.intent:
        print(f"   • {intent}")
    
    print(f"\n❗ Key Issues:")
    for issue in analysis.key_issues:
        print(f"   • {issue}")
    
    print(f"\n🚨 Escalation Needed: {'YES' if analysis.escalation_needed else 'NO'}")
    print(f"⏰ Est. Resolution Time: {analysis.estimated_resolution_time}")
    
    # Test response generation
    print(f"\n📝 Generating Response Suggestion...")
    response_data = agent.generate_response_suggestion(analysis, email_content)
    
    print(f"\n📧 SUGGESTED RESPONSE:")
    print("=" * 50)
    print(response_data["suggested_response"])
    
    print(f"\n🎨 Tone Guidance: {response_data['tone_guidance']}")
    print(f"\n✅ Follow-up Actions: {response_data['follow_up_actions']}")
    
    return True

def test_multiple_scenarios():
    """Test multiple customer scenarios"""
    
    agent = CustomerSuccessCopilot(os.getenv("OPENAI_API_KEY"))
    
    scenarios_to_test = ["billing_question", "positive_feedback", "technical_integration"]
    
    print(f"\n🧪 Testing Multiple Scenarios...")
    
    for scenario in scenarios_to_test:
        print(f"\n" + "="*60)
        print(f"📧 Testing: {scenario}")
        
        test_case = SAMPLE_EMAILS[scenario]
        email_content = test_case["email"]
        customer_profile = test_case.get("customer_profile")
        
        analysis = agent.analyze_customer_communication(email_content, customer_profile)
        
        print(f"   Sentiment: {analysis.sentiment['label']}")
        print(f"   Urgency: {analysis.urgency['level']}")
        print(f"   Intents: {', '.join(analysis.intent)}")
        print(f"   Escalation: {'Yes' if analysis.escalation_needed else 'No'}")

def demo_interview_readiness():
    """Simulate what you'll show in the interview"""
    
    print(f"\n🎭 INTERVIEW DEMO SIMULATION")
    print("=" * 60)
    print("This is what you'll demonstrate to Parag...")
    
    agent = CustomerSuccessCopilot(os.getenv("OPENAI_API_KEY"))
    
    # Use the most impressive scenario
    demo_case = SAMPLE_EMAILS["frustrated_premium"]
    
    print(f"\n📧 Customer Email Preview:")
    print(demo_case["email"][:200] + "...")
    
    print(f"\n⚡ Processing with AI...")
    analysis = agent.analyze_customer_communication(
        demo_case["email"], 
        demo_case["customer_profile"]
    )
    
    print(f"\n🎯 Key Demo Points:")
    print(f"✅ Detected {analysis.sentiment['label']} sentiment with {analysis.sentiment['confidence']:.0%} confidence")
    print(f"✅ Identified {analysis.urgency['level']} urgency requiring escalation")
    print(f"✅ Classified intents: {', '.join(analysis.intent)}")
    print(f"✅ Extracted {len(analysis.key_issues)} specific actionable issues")
    print(f"✅ Premium customer context automatically applied")
    print(f"✅ Generated professional response with appropriate tone")
    
    print(f"\n💰 Business Impact Demo:")
    print(f"✅ Processed complex email in < 5 seconds")
    print(f"✅ Identified escalation need automatically") 
    print(f"✅ Provided structured analysis vs. manual review")
    print(f"✅ Suggested response ready for agent review")
    
    print(f"\n🔧 Technical Sophistication:")
    print(f"✅ Multi-step LLM reasoning pipeline")
    print(f"✅ Customer context integration")
    print(f"✅ Business rule engine (escalation logic)")
    print(f"✅ Structured output for downstream systems")
    
    return True

if __name__ == "__main__":
    print("🚀 Customer Success Copilot - Test Suite")
    print("=" * 50)
    
    # Run comprehensive test
    if test_agent_analysis():
        print(f"\n🎉 Core agent test PASSED!")
        
        test_multiple_scenarios()
        print(f"\n🎉 Multi-scenario test PASSED!")
        
        if demo_interview_readiness():
            print(f"\n🎯 INTERVIEW DEMO READY!")
            print(f"\nNext steps:")
            print(f"✅ Your agent is working perfectly")
            print(f"✅ Ready for W&B Weave integration")
            print(f"✅ Ready for Streamlit demo interface")
            print(f"✅ You have compelling customer scenarios")
            
            print(f"\n💪 Interview Confidence Level: MAXIMUM!")
    else:
        print(f"\n❌ Tests failed - let's debug together!")