requirements.txt
openai
python-dotenv
streamlit
plotly
pandas
weave
wandb